+++ package.json
@@ -40,7 +40,7 @@
     "isomorphic-fetch": "2.2.1",
     "json-loader": "0.5.4",
     "morgan": "^1.7.0",
-    "neo4j-driver": "^1.0.2",
+    "neo4j": "2.0.0-RC2",
     "passport": "^0.3.2",
     "passport-github": "^1.1.0",
     "postcss-loader": "^0.9.1",
