+++ server/models/db.js
@@ -1,6 +1,11 @@
-const neo4j = require('neo4j-driver').v1
-const URL = require('url')
-const url = URL.parse(process.env.GRAPHENEDB_URL)
-console.log(url)
-const driver = neo4j.driver(`bolt://${url.host}`, neo4j.auth.basic(url.auth.split(':')[0], url.auth.split(':')[1]))
-export default driver
+const neo4j = require('neo4j')
+
+const db = new neo4j.GraphDatabase({
+  url: process.env.GRAPHENEDB_URL || 'http://neo4j:neo5j@localhost:7474',
+  // auth: null  // optional; see below for more details
+  // headers: {},    // optional defaults, e.g. User-Agent
+  // proxy: null,    // optional URL
+  // agent: null    // optional http.Agent instance, for custom socket pooling
+})
+
+export default db
